LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/anonymous-iframe/local-storage.tentative.https-expected.txt



PASS Setup
FAIL same_origin anonymous iframe can't access the localStorage assert_equals: expected "" but got "same_origin"
FAIL cross_origin anonymous iframe can't access the localStorage assert_equals: expected "" but got "cross_origin"
